+++ src/LionWeb.Core/Core/M1/NotifyingNodeReplacer.cs
@@ -151,7 +151,7 @@ internal class NotifyingNodeReplacer<T>(INode self, T replacement) : NodeReplace
                 return new ChildMovedAndReplacedFromOtherContainmentInSameParentNotification(_containment, _replacedIndex, replacement, _parent, _oldContainment, _replacementIndex, self,
                     _notificationId);
             case not null when _oldParent != _parent:
-                return new ChildMovedAndReplacedFromOtherContainmentNotification(_parent, _containment, _replacedIndex, replacement, _oldParent, _oldContainment, _replacementIndex, self,
+                return new ChildMovedAndReplacedFromContainmentInOtherParentNotification(_parent, _containment, _replacedIndex, replacement, _oldParent, _oldContainment, _replacementIndex, self,
                     _notificationId);
             default:
                 throw new ArgumentException();